# api/swagger/swagger.yaml
basePath: /api
definitions:
...
  dto.AddDependencyRequest:
    properties:
      blocks_task_id:
        type: string
    required:
    - blocks_task_id
    type: object
...
  dto.CreateTaskTemplateRequest:
    properties:
      assignee_ids:
        items:
          type: string
        type: array
      description:
        type: string
      estimate:
        type: integer
      name:
        maxLength: 200
        minLength: 1
        type: string
      priority:
        type: string
      recurrence_days:
        type: string
      recurrence_pattern:
        type: string
      status_id:
        type: string
    required:
    - name
    - status_id
    type: object
...
  dto.LabelResponse:
    properties:
      color:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
...
  dto.MentionsResponse:
    properties:
      channels:
        additionalProperties:
          type: string
        type: object
      projects:
        additionalProperties:
          type: string
        type: object
      tasks:
        additionalProperties:
          $ref: '#/definitions/dto.TaskMentionResponse'
        type: object
      users:
        additionalProperties:
          type: string
        type: object
    type: object
...
  dto.TaskAssigneeResponse:
    properties:
      avatar_url:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
...
  dto.TaskMentionResponse:
    properties:
      project_id:
        type: string
      title:
        type: string
    type: object
  dto.TaskResponse:
    properties:
      assignees:
        items:
          $ref: '#/definitions/dto.TaskAssigneeResponse'
        type: array
      completed_at:
        type: string
      completed_subtask_count:
        type: integer
      created_at:
        type: string
      created_by:
        type: string
      cycle_id:
        type: string
      description:
        type: string
      due_at:
        type: string
      estimate:
        type: integer
      id:
        type: string
      labels:
        items:
          $ref: '#/definitions/dto.LabelResponse'
        type: array
      mentions:
        allOf:
        - $ref: '#/definitions/dto.MentionsResponse'
        description: |-
          Mentions holds resolved labels for <@type:id> tokens in Description,
          mirroring the chat/comment shape so the frontend reuses the same
          markdown renderer + mention chip rendering.
      parent_task_id:
        type: string
      parent_title:
        type: string
      position_key:
        type: string
      priority:
        type: string
      project_id:
        type: string
      started_at:
        type: string
      status_id:
        type: string
      subtask_count:
        description: |-
          SubtaskCount / CompletedSubtaskCount let the frontend render progress
          (done/total) without client-side counting from a filtered task list.
        type: integer
      subtask_position:
        type: string
      template_id:
        type: string
      title:
        type: string
      updated_at:
        type: string
    type: object
...
  dto.TaskTemplateResponse:
    properties:
      assignee_ids:
        items:
          type: string
        type: array
      created_at:
        type: string
      created_by:
        type: string
      description:
        type: string
      estimate:
        type: integer
      id:
        type: string
      name:
        type: string
      next_run_at:
        type: string
      priority:
        type: string
      project_id:
        type: string
      recurrence_days:
        type: string
      recurrence_pattern:
        type: string
      status_id:
        type: string
      updated_at:
        type: string
    type: object
...
  dto.UpdateTaskTemplateRequest:
    properties:
      assignee_ids:
        items:
          type: string
        type: array
      description:
        type: string
      estimate:
        type: integer
      name:
        maxLength: 200
        minLength: 1
        type: string
      priority:
        type: string
      recurrence_days:
        type: string
      recurrence_pattern:
        type: string
      status_id:
        type: string
    required:
    - name
    - status_id
    type: object
...
  transport.ErrorResponse:
    properties:
      error:
        properties:
          code:
            type: string
          message:
            type: string
        type: object
    type: object
  transport.SuccessResponse:
    properties:
      success:
        type: boolean
    type: object
info:
  contact: {}
  description: Self-hosted project management tool API
  title: Plume API
  version: "1.0"
paths:
...
  /projects/{id}/tasks/{taskId}/dependencies:
    post:
      consumes:
      - application/json
      description: Records that the task is blocked by the given blocking task
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Task ID
        in: path
        name: taskId
        required: true
        type: string
      - description: Blocking task
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.AddDependencyRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/transport.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/transport.ErrorResponse'
      summary: Add a blocking dependency
      tags:
      - task-dependencies
  /projects/{id}/tasks/{taskId}/dependencies/{blocksTaskId}:
    delete:
      description: Removes the blocking edge between two tasks
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Task ID
        in: path
        name: taskId
        required: true
        type: string
      - description: Blocking task ID
        in: path
        name: blocksTaskId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      summary: Remove a blocking dependency
      tags:
      - task-dependencies
  /projects/{id}/tasks/{taskId}/dependencies/blocked:
    get:
      description: Returns the tasks the given task is blocking (these wait on the
        task)
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Task ID
        in: path
        name: taskId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.TaskResponse'
            type: array
      summary: List blocked tasks
      tags:
      - task-dependencies
  /projects/{id}/tasks/{taskId}/dependencies/blocking:
    get:
      description: Returns the tasks blocking the given task (the task is blocked
        by these)
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Task ID
        in: path
        name: taskId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.TaskResponse'
            type: array
      summary: List blocking tasks
      tags:
      - task-dependencies
...
  /projects/{id}/templates:
    get:
      description: Returns all task templates for a project
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.TaskTemplateResponse'
            type: array
      summary: List task templates
      tags:
      - task-templates
    post:
      consumes:
      - application/json
      description: Creates a new task template with optional recurrence
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Template details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateTaskTemplateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.TaskTemplateResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/transport.ErrorResponse'
      summary: Create a task template
      tags:
      - task-templates
  /projects/{id}/templates/{templateId}:
    delete:
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Template ID
        in: path
        name: templateId
        required: true
        type: string
      responses:
        "204":
          description: No Content
      summary: Delete a task template
      tags:
      - task-templates
    patch:
      consumes:
      - application/json
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Template ID
        in: path
        name: templateId
        required: true
        type: string
      - description: Template updates
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateTaskTemplateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TaskTemplateResponse'
      summary: Update a task template
      tags:
      - task-templates
  /projects/{id}/templates/{templateId}/instantiate:
    post:
      description: Creates a real task from a template. For recurring templates, advances
        next_run_at.
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Template ID
        in: path
        name: templateId
        required: true
        type: string
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.TaskResponse'
      summary: Instantiate a task from a template
      tags:
      - task-templates
...
swagger: "2.0"